Split helper into multiple util files
