+++ src/extension/features/accounts/change-enter-behavior/index.js
@@ -1,9 +1,9 @@
 import { Feature } from 'toolkit/extension/features/feature';
-import * as toolkitHelper from 'toolkit/extension/helpers/toolkit';
+import { getCurrentRouteName } from 'toolkit/extension/utils/ynab';
 
 export class ChangeEnterBehavior extends Feature {
   shouldInvoke() {
-    return toolkitHelper.getCurrentRouteName().indexOf('account') !== -1 &&
+    return getCurrentRouteName().indexOf('account') !== -1 &&
            $('.ynab-grid-body-row.is-adding').length;
   }
