+++ src/extension/features/accounts/clear-selection/index.js
@@ -1,12 +1,13 @@
 import { Feature } from 'toolkit/extension/features/feature';
+import { controllerLookup, getEmberView } from 'toolkit/extension/utils/ember';
 
 export class ClearSelection extends Feature {
   uncheckTransactions = () => {
-    let accountsController = ynabToolKit.shared.containerLookup('controller:accounts');
+    let accountsController = controllerLookup('accounts');
 
     try {
       accountsController.get('areChecked').setEach('isChecked', 0);
-      let gridHeader = ynabToolKit.shared.getEmberView($('.ynab-grid-header').attr('id'));
+      let gridHeader = getEmberView($('.ynab-grid-header').attr('id'));
       gridHeader.childViews[0].set('isChecked', false);
       accountsController.send('closeModal');
     } catch (e) {
