+++ src/extension/features/accounts/split-transaction-auto-adjust/index.js
@@ -1,5 +1,5 @@
 import { Feature } from 'toolkit/extension/features/feature';
-import { getCurrentRouteName } from 'toolkit/extension/helpers/toolkit';
+import { getCurrentRouteName } from 'toolkit/extension/utils/ynab';
 
 export class SplitTransactionAutoAdjust extends Feature {
   addAnotherSplit;
