+++ src/extension/features/budget/remove-zero-categories/index.js
@@ -1,5 +1,5 @@
 import { Feature } from 'toolkit/extension/features/feature';
-import { getCurrentRouteName } from 'toolkit/extension/helpers/toolkit';
+import { getCurrentRouteName } from 'toolkit/extension/utils/ynab';
 
 export class RemoveZeroCategories extends Feature {
   shouldInvoke() {
