+++ src/extension/features/budget/target-balance-warning/index.js
@@ -1,5 +1,6 @@
 import { Feature } from 'toolkit/extension/features/feature';
-import * as toolkitHelper from 'toolkit/extension/helpers/toolkit';
+import { getCurrentRouteName } from 'toolkit/extension/utils/ynab';
+import { getEmberView } from 'toolkit/extension/utils/ember';
 
 export class TargetBalanceWarning extends Feature {
   constructor() {
@@ -7,13 +8,13 @@ export class TargetBalanceWarning extends Feature {
   }
 
   shouldInvoke() {
-    return toolkitHelper.getCurrentRouteName().indexOf('budget') !== -1;
+    return getCurrentRouteName().indexOf('budget') !== -1;
   }
 
   invoke() {
     $('.budget-table-row.is-sub-category').each((index, element) => {
       const emberId = element.id;
-      const viewData = toolkitHelper.getEmberView(emberId).data;
+      const viewData = getEmberView(emberId).data;
       const { subCategory } = viewData;
 
       if (subCategory.get('goalType') === ynab.constants.SubCategoryGoalType.TargetBalance) {
