+++ src/extension/features/reports/compact-income-vs-expense/index.js
@@ -1,5 +1,5 @@
 import { Feature } from 'toolkit/extension/features/feature';
-import { getCurrentRouteName } from 'toolkit/extension/helpers/toolkit';
+import { getCurrentRouteName } from 'toolkit/extension/utils/ynab';
 
 export class CompactIncomeVsExpense extends Feature {
   injectCSS() { return require('./index.css'); }
