+++ src/extension/utils/date.js
@@ -1,0 +1,11 @@
+export function getCurrentDate(format) {
+  return ynabDate(format, false);
+}
+
+function ynabDate(format) {
+  if (typeof format !== 'string') {
+    return ynab.YNABSharedLib.dateFormatter.formatDate();
+  }
+
+  return ynab.YNABSharedLib.dateFormatter.formatDate(moment(), format);
+}
